unix connect/bind: compute addr size the same regardless of extra fields
I.e. taking a Unix sockaddr's maximum size, subtracting the one variable
field's max size and adding the size of that field's content gets us the
actual size anytime. This gives us less duplication and only a single
piece of code used across all Unices: i.e. the same code gets tested on
all.

In particular the current makefile, when used on FreeBSD (which *does*
have the sun_len field, but appears to ignore its contents) the
addr-length passed to connect() and bind() is one to small because the
sun_len field's space is neglected in the previous size computation.
This change removes the unnecessary dependency on the build system for
the correctness of the addr length computation.

Signed-off-by: Giel van Schijndel <me@mortis.eu>