Support table errors in socket.newtry/protect
Instead of simply wrapping errors in a table in newtry
and considering all tables exceptions in protect, add
a metatable to exception wrapper and check that. This allows
using protect with functions that may throw error objects.

Additionally, assign __tostring metamethod to the exception
metatable so that unhandled exceptions can be viewed normally.