fix(unix): Pass correct path length for abstract sockets
Since they start with NUL byte, strlen() does the wrong thing.
The Lua API can already pass the real string length, so passing that on
seems the sensible thing to do.

Fixes #216

Removes SUN_LEN macro in favor of the approach used in unixstream.c
since it does not depend on strlen(), which would not work with
NUL-prefixed abstract stream socket paths.